Changed directory order. Exctracted led from glass. Simplified bartender class. Removed nested methods

+++ src/board/Leds.cpp
@@ -1,6 +1,6 @@
-#include "MyLeds.h"
+#include "board/Leds.h"
 
-MyLeds::MyLeds(unsigned int pixels) : Adafruit_NeoPixel(pixels, LEDS_PIN, LEDS_TYPE) {
+MyLeds::MyLeds(uint8_t pixels) : Adafruit_NeoPixel(pixels, LEDS_PIN, LEDS_TYPE) {
     begin();            // Initialize the NeoPixel library
     setBrightness(200); // Set maximum brightness
 }
@@ -12,17 +12,17 @@ void MyLeds::ResetAll(void){
     show();
 }
 
-void MyLeds::ResetGlass(unsigned int num){
+void MyLeds::ResetGlass(uint8_t num){
     SetGlass(num, _black);
     show();
 }
 
-void MyLeds::SetGlass(unsigned int num, uint32_t color){
+void MyLeds::SetGlass(uint8_t num, uint32_t color){
     setPixelColor(num, color);
     show();
 }
 
-void MyLeds::SetGlassPercent(unsigned int num, int percent){
+void MyLeds::SetGlassPercent(uint8_t num, int percent){
     // Calculate red and green intensity based on percentage
     uint8_t red   = (255 * (100 - percent)) / 100;  // Red decreases as percent increases
     uint8_t green = (255 * percent) / 100;          // Green increases as percent increases
